# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PharmaSafe Intelligence API",
    "description": "AI-Powered Pharmaceutical Reverse Chain & Medicine Safety Platform Backend Gateway",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/batches": {
      "post": {
        "tags": [
          "Batches & Custody"
        ],
        "summary": "Create Batch",
        "operationId": "create_batch_api_v1_batches_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Batches & Custody"
        ],
        "summary": "List Batches",
        "operationId": "list_batches_api_v1_batches_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BatchStatusEnum"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "medicine_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Medicine Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BatchResponse"
                  },
                  "title": "Response List Batches Api V1 Batches Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/dead-batches": {
      "get": {
        "tags": [
          "Dead Batch Registry"
        ],
        "summary": "List Dead Batches",
        "operationId": "list_dead_batches_api_v1_dead_batches_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeadBatchResponse"
                  },
                  "title": "Response List Dead Batches Api V1 Dead Batches Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dead-batches/alerts/re-entry-incidents": {
      "get": {
        "tags": [
          "Dead Batch Registry"
        ],
        "summary": "List Reentry Incidents",
        "operationId": "list_reentry_incidents_api_v1_dead_batches_alerts_re_entry_incidents_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VerificationScanResponse"
                  },
                  "title": "Response List Reentry Incidents Api V1 Dead Batches Alerts Re Entry Incidents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dead-batches/{batch_number}": {
      "get": {
        "tags": [
          "Dead Batch Registry"
        ],
        "summary": "Get Dead Batch By Number",
        "operationId": "get_dead_batch_by_number_api_v1_dead_batches__batch_number__get",
        "parameters": [
          {
            "name": "batch_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeadBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchCreate": {
        "properties": {
          "batch_number": {
            "type": "string",
            "title": "Batch Number",
            "example": "AMX-2026-9901"
          },
          "gtin_barcode": {
            "type": "string",
            "title": "Gtin Barcode",
            "example": "08901234567890"
          },
          "medicine_id": {
            "type": "string",
            "title": "Medicine Id"
          },
          "mfg_date": {
            "type": "string",
            "format": "date",
            "title": "Mfg Date"
          },
          "expiry_date": {
            "type": "string",
            "format": "date",
            "title": "Expiry Date"
          },
          "initial_quantity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Initial Quantity"
          },
          "unit": {
            "$ref": "#/components/schemas/UnitEnum",
            "default": "BOX"
          },
          "qr_payload": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qr Payload"
          }
        },
        "type": "object",
        "required": [
          "batch_number",
          "gtin_barcode",
          "medicine_id",
          "mfg_date",
          "expiry_date",
          "initial_quantity"
        ],
        "title": "BatchCreate"
      },
...
      "BatchResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "batch_number": {
            "type": "string",
            "title": "Batch Number"
          },
          "gtin_barcode": {
            "type": "string",
            "title": "Gtin Barcode"
          },
          "medicine_id": {
            "type": "string",
            "title": "Medicine Id"
          },
          "manufacturer_id": {
            "type": "string",
            "title": "Manufacturer Id"
          },
          "mfg_date": {
            "type": "string",
            "format": "date",
            "title": "Mfg Date"
          },
          "expiry_date": {
            "type": "string",
            "format": "date",
            "title": "Expiry Date"
          },
          "initial_quantity": {
            "type": "integer",
            "title": "Initial Quantity"
          },
          "current_quantity": {
            "type": "integer",
            "title": "Current Quantity"
          },
          "unit": {
            "$ref": "#/components/schemas/UnitEnum"
          },
          "status": {
            "$ref": "#/components/schemas/BatchStatusEnum"
          },
          "current_custodian_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Custodian Id"
          },
          "is_recalled": {
            "type": "boolean",
            "title": "Is Recalled"
          },
          "recall_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recall Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "batch_number",
          "gtin_barcode",
          "medicine_id",
          "manufacturer_id",
          "mfg_date",
          "expiry_date",
          "initial_quantity",
          "current_quantity",
          "unit",
          "status",
          "is_recalled",
          "created_at",
          "updated_at"
        ],
        "title": "BatchResponse"
      },
      "BatchStatusEnum": {
        "type": "string",
        "enum": [
          "MANUFACTURED",
          "IN_DISTRIBUTION",
          "AT_PHARMACY",
          "DISPENSED",
          "EXPIRED",
          "RECALLED",
          "FLAGGED_SUSPICIOUS",
          "RETURN_INITIATED",
          "RETURN_IN_TRANSIT",
          "RECEIVED_AT_DISPOSAL",
          "DESTROYED",
          "DEAD_BATCH"
        ],
        "title": "BatchStatusEnum"
      },
...
      "DeadBatchResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "batch_number": {
            "type": "string",
            "title": "Batch Number"
          },
          "gtin_barcode": {
            "type": "string",
            "title": "Gtin Barcode"
          },
          "destruction_record_id": {
            "type": "string",
            "title": "Destruction Record Id"
          },
          "destruction_cert_hash": {
            "type": "string",
            "title": "Destruction Cert Hash"
          },
          "manufacturer_name": {
            "type": "string",
            "title": "Manufacturer Name"
          },
          "quantity_destroyed": {
            "type": "integer",
            "title": "Quantity Destroyed"
          },
          "destroyed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Destroyed At"
          },
          "blacklisted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Blacklisted At"
          },
          "reentry_attempts_count": {
            "type": "integer",
            "title": "Reentry Attempts Count"
          },
          "last_reentry_detected_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Reentry Detected At"
          },
          "is_actively_monitored": {
            "type": "boolean",
            "title": "Is Actively Monitored"
          }
        },
        "type": "object",
        "required": [
          "id",
          "batch_id",
          "batch_number",
          "gtin_barcode",
          "destruction_record_id",
          "destruction_cert_hash",
          "manufacturer_name",
          "quantity_destroyed",
          "destroyed_at",
          "blacklisted_at",
          "reentry_attempts_count",
          "is_actively_monitored"
        ],
        "title": "DeadBatchResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UnitEnum": {
        "type": "string",
        "enum": [
          "TABLET_STRIP",
          "BOTTLE",
          "VIAL",
          "BOX",
          "AMPOULE"
        ],
        "title": "UnitEnum"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VerificationScanResponse": {
        "properties": {
          "verification_status": {
            "$ref": "#/components/schemas/VerificationStatusEnum"
          },
          "batch_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Number"
          },
          "brand_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Name"
          },
          "generic_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generic Name"
          },
          "manufacturer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manufacturer Name"
          },
          "is_expired": {
            "type": "boolean",
            "title": "Is Expired",
            "default": false
          },
          "is_recalled": {
            "type": "boolean",
            "title": "Is Recalled",
            "default": false
          },
          "is_dead_batch_reentry": {
            "type": "boolean",
            "title": "Is Dead Batch Reentry",
            "default": false
          },
          "warning_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warning Message"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "scan_id": {
            "type": "string",
            "title": "Scan Id"
          }
        },
        "type": "object",
        "required": [
          "verification_status",
          "timestamp",
          "scan_id"
        ],
        "title": "VerificationScanResponse"
      },
      "VerificationStatusEnum": {
        "type": "string",
        "enum": [
          "AUTHENTIC",
          "EXPIRED",
          "RECALLED",
          "FLAGGED_SUSPICIOUS",
          "DEAD_BATCH_REENTRY_DETECTED",
          "UNKNOWN_NOT_FOUND"
        ],
        "title": "VerificationStatusEnum"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}